+++ wsgidav/debug_filter.py
@@ -62,6 +62,7 @@ import threading
 
 from wsgidav import compat, util
 from wsgidav.middleware import BaseMiddleware
+from wsgidav.util import safeReEncode
 
 __docformat__ = "reStructuredText"
 
@@ -146,7 +147,7 @@ class WsgiDavDebugFilter(BaseMiddleware):
                   (threading.currentThread().ident, method), file=self.out)
             for k, v in environ.items():
                 if k == k.upper():
-                    print("%20s: '%s'" % (k, v), file=self.out)
+                    print("%20s: '%s'" % (k, safeReEncode(v, self.out.encoding)), file=self.out)
             print("\n", file=self.out)
 
         # Intercept start_response